refactor: 폴더 구조변경 refactor: response 코드 짧게 수정
